+++ docs/CHANGELOG.rst
@@ -38,6 +38,7 @@ Fixed
 - Fix docker container name for parallel tests
 - Generate temporary names for upload files in tests
 - Fix permissions in Elasticsearch tests
+- Do not purge data in tests
 
 
 ==================